+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/DTO/PricingGuide/OpenMarketTotalsDto.cs
@@ -2,7 +2,7 @@
 {
     public class OpenMarketTotalsDto
     {
-        public double Cpm { get; set; }
+        public decimal Cpm { get; set; }
         public decimal Cost { get; set; }
         public double Impressions { get; set; }
         public double Coverage { get; set; }
